@@ -1,6 +1,6 @@
 {
   "name": "ngx-summernote-app",
-  "version": "1.0.0",
+  "version": "18.0.0",
   "keywords": [
     "summernote",
     "angular"
@@ -11,46 +11,44 @@
     "build": "ng build",
     "build:lib": "ng build --configuration production ngx-summernote && cp README.md LICENSE dist/ngx-summernote",
     "publish": "npm run build:lib && cd dist/ngx-summernote && npm publish",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "lint": "ng lint"
   },
   "dependencies": {
-    "@angular/animations": "^14.3.0",
-    "@angular/common": "^14.3.0",
-    "@angular/compiler": "^14.3.0",
-    "@angular/core": "^14.3.0",
-    "@angular/forms": "^14.3.0",
-    "@angular/platform-browser": "^14.3.0",
-    "@angular/platform-browser-dynamic": "^14.3.0",
-    "@angular/router": "^14.3.0",
+    "@angular/animations": "^18.2.11",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/forms": "^18.2.11",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11",
     "core-js": "^3.5.0",
     "jquery": "^3.5.0",
     "rxjs": "^6.5.3",
-    "summernote": "^0.8.18",
+    "summernote": "^0.9.1",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.2.11",
-    "@angular/cli": "^14.2.11",
-    "@angular/compiler-cli": "^14.3.0",
-    "@angular/language-service": "^14.3.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.8",
+    "@angular-devkit/build-angular": "18.2.11",
+    "@angular-eslint/builder": "18.2.0",
+    "@angular-eslint/eslint-plugin": "18.2.0",
+    "@angular-eslint/eslint-plugin-template": "18.2.0",
+    "@angular-eslint/schematics": "18.2.0",
+    "@angular-eslint/template-parser": "18.2.0",
+    "@angular/cli": "18.2.11",
+    "@angular/compiler-cli": "^18.2.11",
+    "@angular/language-service": "^18.2.11",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "^3.99.1",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.4.2",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^14.2.2",
-    "protractor": "~7.0.0",
+    "@typescript-eslint/eslint-plugin": "8.13.0",
+    "@typescript-eslint/parser": "8.13.0",
+    "angular-eslint": "18.4.0",
+    "codelyzer": "^6.0.2",
+    "eslint": "^9.14.0",
+    "husky": "9.1.6",
+    "ng-packagr": "^18.2.1",
     "ts-node": "~8.5.4",
-    "tslint": "~6.1.0",
-    "typescript": "~4.8.0"
+    "typescript": "~5.4.5",
+    "typescript-eslint": "8.10.0"
   }
 }
