@@ -16,9 +16,9 @@
     "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/common": "14.x - 16.x",
-    "@angular/core": "14.x - 16.x",
-    "@angular/forms": "14.x - 16.x",
-    "summernote": ">=0.8.10"
+    "@angular/common": "18.x",
+    "@angular/core": "18.x",
+    "@angular/forms": "18.x",
+    "summernote": ">=0.9.1"
   }
 }
